ai_voice_call/requirements.txt
# Voice call specific dependencies
fastapi
uvicorn
python-dotenv

# Google AI services
google-generativeai

# Audio processing
soundfile
librosa

# Flask-based voice app dependencies
flask
flask-sock
google-cloud-aiplatform

# WebSocket and async support
gevent
geventwebsocket
